Child to Parent: Sharing Data via ViewChild
Child to Parent: Sharing Data via ViewChild : ViewChild allows a one component to be injected into another, giving the parent access to its attributes and functions. One caveat, however, is that child won’t be available until after the view has been initialized. This means we need to implement the AfterViewInit lifecycle hook to receive the data from the child.